(Feat) play tetris code

+++ src/HAL/Timers/timer1/timer1.h
@@ -73,6 +73,6 @@ typedef struct {
 
 void initTimer1(const s_TIM1_CONFIG *config);
 
-void setOCR1A(uint8_t value);
+void setOCR1A(uint16_t value);
 
 #endif //TIMER1_H
